refactor: pure pytest tests instead of unittest
